Renamed ExamplePlugin to ExamplePdfPlugin (to fall in line with the naming of the HTML5 and, eventually, the AdaptiveForm plugins).
